Added Utils file
Contains common functions such as saveJson, and cleanHtml
